# docs/swagger.yaml
definitions:
  helper.HTTPError:
    properties:
      code:
        example: 999
        type: integer
      message:
        example: example response
        type: string
    type: object
...
  model.Users:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /user/create:
    post:
      consumes:
      - application/json
      description: Create registers by record
      parameters:
      - description: Create supplier
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.Users'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Users'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.HTTPError'
        "404":
          description: A User ID with the specified ID was not found
          schema:
            $ref: '#/definitions/helper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.HTTPError'
      summary: Register user
      tags:
      - users
  /user/get/all:
    get:
      consumes:
      - application/json
      description: get user information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Users'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.HTTPError'
        "404":
          description: A Gender was not found
          schema:
            $ref: '#/definitions/helper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.HTTPError'
      summary: Get all user data
      tags:
      - users
swagger: "2.0"